- Modified the currentQuestion property from type int to Long. - In QuizService, modified the selectedOptions property from type String to Long. - Created processAnswer() method in QuizService. - Created Skeleton for PatchMapping route in QuizController.

+++ src/main/java/com/example/the_best_quiz_ever/models/Quiz.java
@@ -18,7 +18,7 @@ public class Quiz {
     @Column
     private String name;
     @Column
-    private int currentQuestion;
+    private Long currentQuestion;
 
     @OneToMany(mappedBy = "quiz")
     @JsonIgnoreProperties({"quiz"})
@@ -29,7 +29,7 @@ public class Quiz {
 //    private List<Outcome> outcomes;
 
     //    constructor
-    public Quiz(String name, int currentQuestion) {
+    public Quiz(String name, Long currentQuestion) {
         this.name = name;
 //        this.outcomes = new ArrayList<>();
         this.currentQuestion = currentQuestion;
@@ -81,11 +81,11 @@ public class Quiz {
 //        this.outcomes = outcomes;
 //    }
 
-    public int getCurrentQuestion() {
+    public Long getCurrentQuestion() {
         return currentQuestion;
     }
 
-    public void setCurrentQuestion(int currentQuestion) {
+    public void setCurrentQuestion(Long currentQuestion) {
         this.currentQuestion = currentQuestion;
     }
